move runner file
